gateway: refactor package gateway.headers as gateway.filter.headers
