+++ gateway/src/main/java/org/georchestra/gateway/filter/headers/HeaderProvider.java
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU General Public License along with
  * geOrchestra.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.georchestra.gateway.headers;
+package org.georchestra.gateway.filter.headers;
 
 import java.util.function.Consumer;
